fix(Template): Add special constant for compiler directive
When the template uses #if, then the compiler interprets those when
installing the template. However, the template defines directives that
should be used in the actual app so we needs the template generator to
ignore those directives, but "activate them" with `dotnet new`. This can
be accomplished with the constant macro and commenting out the directive

fixes AB#4596, #16