feat(uow): Add method to find an aggregate root
Useful in event handling you an event emits with an aggregate root id.
You should safely assume it is already in the unit of work so you do not
have to go looking in the repository. Looking in the repository is a bit
awkward because you do not know if it is hitting a database or the
repository has an identity map implementation

fixes AB#5628