fix(query): Do not convert Contains operators in query
Do not convert an `IEnumerable` in the query to the contract data type.
This will throw a coercion error because it will try to convert, for
example, and `IEnumerable<int>` to an `int`. Queries can map an
`IEnuemrable` search property to a contract property `int` because it
create a `Contains` search such as `in (1,2,3)`. The purpose of this
conversion is to make data types such as DateTime? and DateTime the same.