fix(Events): Change CorrelationId to 1 to many with Metadata
* The correlation id ties together commands that create events and other
events. Therefore, the relationship to the metadata through the
correlation id is a one to many because many events can be correlated to
one metadata
* Fix metadata event tracking. Multiple events tied to the same request
was creating extra metadata records rather than relating it to the same
metadata already in the `DbContext`. Check to see if metadata
exists before adding a new event metadata to keep the relationship
* test(migrations): Update the `DomainEventStub` name to `EventStub` from
commit f7ff8b6ee7bdcc294548c6c2162b6c4a29e2d73b

merge pr  (#133)