fix(json): Fix error handling when payload is invalid json
* Make sure to rewind bodyready so the request can be re-read, even on an
error. I was seeing weird errors from Kestrel that the response was
being read as a last exception, even though this exception was
swallowed. Do not complete the reading with an exception since it always
returns a 500. Most times we want to return a 422 if the payload is invalid
during deserialization.
* Logger should return original exception so response decorator can do
proper mapping in the problem details factory. Using the app exception
will result in a 500 and hide the actual error that could be mapped to a
status code

Other Boy scout JSON stuff:
* Correct System.Json file names that do not being with System.Json.
  Those were not filtered out by the template
* Remove custom error handling in newtonsoft, it really is pointless and
we do not want to make to a model exception that translates to a 404 if the
entity is unprocessable.
* Add missing tests
* Add app specific exceptions so we can tell the different between a
serialization vs deserialization error.
* Remove System & Newtonsoft registration. Exception decorators are no 
longer needed. Mapping to a status code is needed in the problem detail
options used in the generic Json Registration

merge pr (#136)