fix(efcore): Add pre-compiled discriminator
When no discriminator value is specified, it used the name of the class
(not the full type name). However, we expect it to be the dataset name,
which is the full type name of the class, minus the assembly info. In
order to set this value, we set it in the entity configuration class. If
we do not do this, then our efcore will use the name of the class in the
`where` clause of the query, which will be the name of the class and not
match our inserted value of the type name of the class