fix(efcore): Require discriminator value on generic `Metadata`
Require a value for the discriminator value in the entity configuration
or else it will default to the name of the class, which will not match
the `ToString()` value used at runtime for the discriminator column.
That does not work because EFCore uses a const value given at setup time
and places this in the `where` clause. Therefore, create a new column
for discriminator so if we query the non generic `Metadata` class, we
can still get the different types that were inserted by using the
`DataSetName` name.