refactor(db): Remove transaction factory
* Not needed most of the time & adds overhead. Can be solved a better way
with another decorator rather than in the current decorator
* Add async extensions to run actions that do not return anything, but just pass
the original result on. Good for running side effect function. Make sure the name
is different so we do not confuse it with `AndThen`, which returns something. This
will make the current commit/revert logic flow better in the transaction decorator.

merge PR (#146)